WIP refactor to move github and feature branch manager to separate CDK projects and reuse config management strategy
