+++ infrastructure/requirements.txt
@@ -5,3 +5,4 @@ aws-cron-expression-validator==1.0.4
 checksumdir==1.2.0
 typer
 pydantic==1.10.17
+../utilities/packages/python/provena-shared-functionality
