+++ utilities/supporting-stacks/feature-branch-manager/.gitignore
@@ -1,0 +1,10 @@
+*.swp
+package-lock.json
+__pycache__
+.pytest_cache
+.venv
+*.egg-info
+
+# CDK asset staging directory
+.cdk.staging
+cdk.out
