+++ utilities/supporting-stacks/feature-branch-manager/feature_branch_manager/feature_branch_manager_stack.py
@@ -1,0 +1,19 @@
+from aws_cdk import (
+    # Duration,
+    Stack,
+    # aws_sqs as sqs,
+)
+from constructs import Construct
+
+class FeatureBranchManagerStack(Stack):
+
+    def __init__(self, scope: Construct, construct_id: str, **kwargs) -> None:
+        super().__init__(scope, construct_id, **kwargs)
+
+        # The code that defines your stack goes here
+
+        # example resource
+        # queue = sqs.Queue(
+        #     self, "FeatureBranchManagerQueue",
+        #     visibility_timeout=Duration.seconds(300),
+        # )
