+++ utilities/supporting-stacks/feature-branch-manager/requirements-dev.txt
@@ -1,0 +1,1 @@
+pytest==6.2.5
