+++ utilities/supporting-stacks/github-creds/configs/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!sample.json
