+++ utilities/supporting-stacks/github-creds/configs/sample.json
@@ -1,0 +1,7 @@
+{
+    "env": {
+        "account": "239021088610",
+        "region": "ap-southeast-2"
+    },
+    "github_token_arn": ""
+}
