+++ utilities/supporting-stacks/github-creds/mypy.ini
@@ -1,0 +1,13 @@
+[mypy]
+exclude = (?x)(
+    ^cdk\.out   # or files starting with "three."
+    | ^\.venv.*
+    | .*\.out
+    | .*__pycache__.*
+    | .*pytest_cache.*
+    | .*lambda_utility.*
+    | .*prov_job_dispatcher.*
+  )
+disallow_untyped_defs = True 
+disallow_incomplete_defs = True
+allow_redefinition = True
